refactor: rewrite the resolve method *again* (wip)
