+++ packages/openapi-parser/src/lib/resolve/resolve.ts
@@ -1,6 +1,4 @@
 import { AnyObject } from '../../types'
-import { traverse } from '../../utils'
-import { isObject } from './isObject'
 import { unescapeJsonPointer } from './unescapeJsonPointer'
 
 // TODO: Add support for all pointer words
@@ -16,139 +14,76 @@ import { unescapeJsonPointer } from './unescapeJsonPointer'
 /**
  * Takes a specification and resolves all references.
  */
-export function resolve(
+export function resolve(input: AnyObject) {
+  // Detach from input
+  const specification = structuredClone(input)
+
+  // Recursively resolve all references
+  resolveReferences(specification)
+
+  // If we replace references with content, that includes a reference, we can’t deal with that right-away.
+  // That’s why we need a second run.
+  resolveReferences(specification)
+
+  // Return the resolved specification
+  return specification
+
   /**
-   * We need the whole specification, otherwise we can’t resolve the references.
+   * Resolves the circular reference to an object and deletes the $ref properties.
    */
-  wholeSpecification: AnyObject,
-  /**
-   * If `true`, the original specification will be modified.
-   * If `false`, the original specification will be left untouched.
-   */
-  replace: boolean = true,
-  /**
-   * Sometimes, we only want to resolve references for a part of the specification.
-   */
-  partialSpecification?: AnyObject,
-) {
-  // Make sure we’re dealing with an object
-  if (!isObject(partialSpecification ?? wholeSpecification)) {
-    return undefined
-  }
+  function resolveReferences(schema: AnyObject) {
+    // Iterate over the whole objecct
+    Object.entries(schema).forEach(([key, value]) => {
+      // Ignore parts without a reference
+      if (schema.$ref !== undefined) {
+        // Find the referenced content
+        const target = resolveUri(specification, schema.$ref)
 
-  let modifiedSpecification: AnyObject
+        if (target === undefined) {
+          throw new Error(`Can’t resolve URI: ${schema.$ref}`)
+        }
 
-  // Write
-  if (replace) {
-    modifiedSpecification = partialSpecification ?? wholeSpecification
-  }
-  // Read-only (detach from original object)
-  else {
-    modifiedSpecification = structuredClone(
-      partialSpecification ?? wholeSpecification,
-    )
-  }
+        // Get rid of the reference
+        delete schema.$ref
 
-  // Go through the whole wholeSpecification and resolve all references
-  return traverse(modifiedSpecification, (schema) => {
-    // Ignore parts without a reference
-    if (schema.$ref === undefined) {
-      return schema
-    }
+        if (typeof target === 'object') {
+          Object.keys(target).forEach((key) => {
+            return (schema[key] = target[key])
+          })
+        }
+      }
 
-    // Oh, great, there’s a reference! Let’s resolve it.
-    const target = resolveUri(wholeSpecification, schema.$ref)
-
-    // If we can’t find the reference, we throw an error.
-    if (target === undefined) {
-      throw new Error(`Can’t resolve URI: ${schema.$ref}`)
-    }
-
-    // If the URI resolves to the partial specification object, we have a circular reference
-    const isCircular = target === modifiedSpecification
-
-    // Get rid of the $ref property
-    delete schema.$ref
-
-    // Before we put the referenced content into place, we should resolve any references inside the reference.
-    // Do not continue the recursion if we have a circular reference
-    const resolvedTarget = isCircular
-      ? resolveCircularReference(target)
-      : resolve(wholeSpecification, replace, target)
-
-    // We want to keep the reference to the original object, but add the original properties:
-    Object.keys(resolvedTarget ?? {}).forEach((key) => {
-      schema[key] = resolvedTarget[key]
+      if (typeof value === 'object' && !isCircular(value)) {
+        resolveReferences(value)
+      }
     })
+  }
+}
 
-    return schema
-  })
+// TODO: Is there a better way? :D
+function isCircular(schema: AnyObject) {
+  try {
+    JSON.stringify(schema)
+    return false
+  } catch (error) {
+    return true
+  }
 }
 
 /**
- * Resolves the circular reference to an object and deletes the $ref properties
+ * Resolves a URI to a part of the specification
  */
-function resolveCircularReference(circularSpec: AnyObject) {
-  /**
-   * Iterate over a nested object recursively and delete all $ref properties
-   */
-  const deleteNestedReferences = (schema: AnyObject) => {
-    Object.keys(schema).forEach((key) => {
-      if (key === '$ref' && schema[key] !== null) {
-        delete schema[key]
-      }
-
-      if (typeof schema[key] === 'object' && schema[key] !== null) {
-        deleteNestedReferences(schema[key])
-      }
-    })
-
-    return schema
-  }
-
-  /**
-   * Iterate over a nested object recursively and replace $ref with the specified element
-   */
-  const addNestedReferences = (schema: AnyObject, element: AnyObject) => {
-    Object.keys(schema).forEach((key) => {
-      if (typeof schema[key] === 'object' && schema[key] !== null) {
-        const nested = schema[key]
-
-        Object.keys(nested).forEach((nestedKey) => {
-          if (nestedKey === '$ref' && schema[nestedKey] !== null) {
-            schema[key] = element
-          }
-        })
-
-        addNestedReferences(schema[key], element)
-      }
-    })
-    return schema
-  }
-
-  // Create the circular reference object by removing the $ref properties
-  const circularRefObj = deleteNestedReferences(structuredClone(circularSpec))
-
-  // Iterate over the circular spec and replace the $ref with the circular reference object
-  return addNestedReferences(structuredClone(circularSpec), circularRefObj)
-}
-
-/**
- * Get the specified URI from a given specification
- */
-function resolveUri(specification: AnyObject, uri: string) {
+function resolveUri(specification: AnyObject, uri: string): AnyObject {
   // Understand the URI
   const [prefix, path] = uri.split('#', 2)
-  // const hasHash = !!path
-  const segments = path.split('/').map(unescapeJsonPointer).slice(1)
 
   if (prefix) {
-    throw new Error(`External references are not supported yet: ${prefix}`)
+    throw new Error(`External references are not supported yet: ${uri}`)
   }
 
-  // Get the target
-  return segments.reduce(
-    (accumulator: AnyObject, segment: string) => accumulator[segment],
-    specification,
-  )
+  const segments = path.split('/').slice(1)
+
+  return segments.reduce((acc, key) => {
+    return acc[key]
+  }, specification)
 }
